+++ src/main/java/micdoodle8/mods/galacticraft/core/perlin/generator/Billowed.java
@@ -8,30 +8,31 @@ import micdoodle8.mods.galacticraft.core.perlin.NoiseModule;
 public class Billowed extends NoiseModule
 {
 	private final FishyNoise noiseGen;
-	private final double offsetX;
-	private final double offsetY;
-	private final double offsetZ;
+	private final float offsetX;
+	private final float offsetY;
+	private final float offsetZ;
 	private final int numOctaves;
-	private final double persistance;
+	private final float persistance;
 
-	public Billowed(long seed, int nOctaves, double p)
+	public Billowed(long seed, int nOctaves, float p)
 	{
 		this.numOctaves = nOctaves;
 		this.persistance = p;
 		final Random rand = new Random(seed);
-		this.offsetX = rand.nextDouble() / 2 + 0.01D;
-		this.offsetY = rand.nextDouble() / 2 + 0.01D;
-		this.offsetZ = rand.nextDouble() / 2 + 0.01D;
+		this.offsetX = rand.nextFloat() / 2 + 0.01F;
+		this.offsetY = rand.nextFloat() / 2 + 0.01F;
+		this.offsetZ = rand.nextFloat() / 2 + 0.01F;
 		this.noiseGen = new FishyNoise(seed);
 	}
 
 	@Override
-	public double getNoise(double i)
+	public float getNoise(float i)
 	{
 		i *= this.frequencyX;
-		double val = 0;
-		double curAmplitude = this.amplitude;
-		for(int n = 0; n < this.numOctaves; n++) {
+		float val = 0;
+		float curAmplitude = this.amplitude;
+		for (int n = 0; n < this.numOctaves; n++)
+		{
 			val += Math.abs(this.noiseGen.noise2d(i + this.offsetX, this.offsetY) * curAmplitude);
 			i *= 2;
 			curAmplitude *= this.persistance;
@@ -40,12 +41,12 @@ public class Billowed extends NoiseModule
 	}
 
 	@Override
-	public double getNoise(double i, double j)
+	public float getNoise(float i, float j)
 	{
 		i *= this.frequencyX;
 		j *= this.frequencyY;
-		double val = 0;
-		double curAmplitude = this.amplitude;
+		float val = 0;
+		float curAmplitude = this.amplitude;
 		for (int n = 0; n < this.numOctaves; n++)
 		{
 			val += Math.abs(this.noiseGen.noise2d(i + this.offsetX, j + this.offsetY) * curAmplitude);
@@ -57,12 +58,12 @@ public class Billowed extends NoiseModule
 	}
 
 	@Override
-	public double getNoise(double i, double j, double k)
+	public float getNoise(float i, float j, float k)
 	{
 		i *= this.frequencyX;
 		j *= this.frequencyY;
 		k *= this.frequencyZ;
-		double val = 0;
+		float val = 0;
 		for (int n = 0; n < this.numOctaves; n++)
 		{
 			val += Math.abs(this.noiseGen.noise3d(i + this.offsetX, j + this.offsetY, k + this.offsetZ) * this.amplitude);
